requirements.txt
telethon~=1.21.1
mysql.connector